Pushing the income back up to 5.5
Did not have the desired effect revenue is down anyway with the drop in bounty from 1.1 to 0.7. Default was 6 before I altered it. Can someone explain to be how income from unoccupied towns works? I tried reading the script and could not determine if empty but fully controlled towns are giving their worth.